refactor - 상태 관련 클래스를 상태 패키지로 이동
