+++ src/test/java/bowling/domain/GutterTest.java
@@ -1,5 +1,6 @@
 package bowling.domain;
 
+import bowling.domain.state.Gutter;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
