+++ src/test/java/bowling/domain/SpareTest.java
@@ -1,5 +1,6 @@
 package bowling.domain;
 
+import bowling.domain.state.Spare;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
